cetlib: fixup rpaths
cetlib supplies a helper executable linked to the main cetlib
library. The build system does not set the install RPATH of the
program to point to the library, so a default install fails.

Use MachO tools to change the install_name of the library used
by the program to its full path, per Homebrew guidelines. Limit
change to macOS, awaiting test on Linuxbrew.

Add test of running program to confirm that it runs o.k.

Bump revision to ensure updates.

TODO: Report issue upstream and review how best to handle these
mixed packages.